refactor: rename materiales module to materials (English)
Rename the entire materiales feature to materials across folders, files,
components, API client, navigation, and i18n keys. Aligns the admin panel
with the parallel backend rename that establishes /materials/* as the
canonical URL contract.

Scope:
- folder renames: dashboard/materiales/ -> dashboard/materials/, with
  subroutes bandeja->inbox, comprobantes->receipts, inventario->inventory,
  configuracion->config, categorias->categories, solicitud->request
- folder src/components/materiales/ -> src/components/materials/
- file renames: bandeja-filters->inbox-filters, categorias-table->categories-table,
  categoria-form-sheet->category-form-sheet, delete-categoria-dialog->
  delete-category-dialog, nueva-categoria-button->new-category-button,
  inventario-filters->inventory-filters, inventario-table->inventory-table,
  nuevo-producto-button->new-product-button, configuracion-form->config-form,
  comprobantes-filters->receipts-filters, comprobante-review-actions->
  receipt-review-actions, comprobantes-section->receipts-section,
  estado-badge->status-badge
- src/lib/api/materiales.ts -> src/lib/api/materials.ts
- src/lib/types/materiales.ts -> src/lib/types/materials.ts
- API client function renames (18): listCategoriasAdmin->listCategoriesAdmin,
  createCategoria->createCategory, listOrdenes->listOrders, aprobarOrden->
  approveOrder, cancelarOrden->cancelOrder, entregarOrden->deliverOrder,
  patchOrdenLinea->patchOrderLine, listInventario->listInventory,
  listCatalogo->listCatalog, getConfiguracion->getConfig,
  listConfiguracionAll->listConfigAll, updateConfiguracion->updateConfig,
  listComprobantes->listReceipts, aprobarComprobante->approveReceipt,
  rechazarComprobante->rejectReceipt
- backend URLs updated to /materials/* with English subpaths
- i18n nav keys nav.sections.materiales -> materials, nav.items.materiales_*
  -> materials_* in en/es/fr/pt-BR
- nav-config.ts updated to new URLs and i18n keys

Unchanged (backend contract):
- Permission strings 'materiales:*' in hasPermission() calls
- User-facing Spanish UI copy in page titles and form descriptions
