+++ messages/en.json
@@ -11,7 +11,7 @@
       "members": "Members",
       "operations": "Operations",
       "rankings_analytics": "Rankings & Analytics",
-      "materiales": "Materials"
+      "materials": "Materials"
     },
     "items": {
       "dashboard": "Dashboard",
@@ -104,10 +104,11 @@
       "subgroup_business": "Business",
       "subgroup_honors_cat": "Honors",
       "ranking_weights_root": "Weights configuration",
-      "materiales_bandeja": "Inbox",
-      "materiales_comprobantes": "Receipts",
-      "materiales_inventario": "Inventory",
-      "materiales_configuracion": "Settings"
+      "materials_inbox": "Inbox",
+      "materials_receipts": "Receipts",
+      "materials_inventory": "Inventory",
+      "materials_categories": "Categories",
+      "materials_config": "Settings"
     },
     "breadcrumbs": {
       "dashboard": "Dashboard",
@@ -2892,16 +2893,126 @@
       },
       "detail": {
         "title": "User detail",
+        "subtitle": "Complete information, access, health, and activity for this user.",
         "back": "Back",
         "restrictedTitle": "Restricted access",
         "restrictedDescription": "You need resolved global authorization to query data for other users.",
         "statusActive": "Active",
         "statusInactive": "Inactive",
+        "ageYears": "{count, plural, one {# year} other {# years}}",
+        "approvalApproved": "Approved",
+        "approvalPending": "Pending",
+        "approvalRejected": "Rejected",
+        "tabResumen": "Summary",
+        "tabPersonalData": "Personal data",
+        "tabHealth": "Health",
+        "tabRolesAccess": "Roles & access",
         "tabInfo": "Information",
         "tabPostRegistration": "Post-registration",
         "tabSecurity": "Security",
         "tabSessions": "Sessions",
-        "postRegistrationUnavailable": "Could not retrieve post-registration status. The user may not have started the process or there may be a connectivity error."
+        "postRegistrationUnavailable": "Could not retrieve post-registration status. The user may not have started the process or there may be a connectivity error.",
+        "tenure": {
+          "years": "{count, plural, one {# year} other {# years}}",
+          "months": "{count, plural, one {# month} other {# months}}",
+          "days": "{count, plural, one {# day} other {# days}}"
+        },
+        "stats": {
+          "tenureLabel": "Time in SACDIA",
+          "tenureRegisteredOn": "Registered on {date}",
+          "classesLabel": "Classes",
+          "classesActive": "Active enrollments",
+          "rolesLabel": "Roles",
+          "rolesEmpty": "No roles",
+          "rolesRbacNote": "RBAC view",
+          "lastUpdateLabel": "Last update",
+          "panelEnabled": "Panel enabled",
+          "panelDisabled": "Panel disabled",
+          "auditTracking": "Audit tracking"
+        },
+        "sidebar": {
+          "title": "Access & status",
+          "status": "Status",
+          "approval": "Approval",
+          "roles": "Roles",
+          "appAccess": "App access",
+          "panelAccess": "Panel access",
+          "mfa": "MFA",
+          "assignments": "Assignments",
+          "enabled": "Enabled",
+          "disabled": "Disabled",
+          "yes": "Yes",
+          "no": "No",
+          "dash": "—"
+        },
+        "sections": {
+          "identityTitle": "Identity",
+          "locationTitle": "Location",
+          "fullDataTitle": "Full data",
+          "rolesAccessTitle": "Roles & access",
+          "scopeTitle": "Scope",
+          "rolesLabel": "Active roles",
+          "rolesEmpty": "This user has no assigned roles.",
+          "contactsTitle": "Emergency contacts",
+          "contactsEmpty": "No emergency contacts registered.",
+          "contactsMissingPayload": "Could not load contacts.",
+          "legalRepTitle": "Legal representative",
+          "legalRepEmpty": "No legal representative registered.",
+          "pastoralTitle": "Pastoral",
+          "pastoralEmpty": "No pastoral information registered.",
+          "healthTitle": "Health",
+          "healthBlockTitle": "Medical information",
+          "healthEmpty": "No medical information registered.",
+          "healthShow": "Show",
+          "healthHide": "Hide",
+          "healthProtectedTitle": "Protected information",
+          "healthProtectedDescription": "Medical data is only shown to roles with explicit permission."
+        },
+        "fields": {
+          "fullName": "Full name",
+          "name": "First name",
+          "paternalLastName": "Last name (paternal)",
+          "maternalLastName": "Last name (maternal)",
+          "birthday": "Date of birth",
+          "gender": "Gender",
+          "baptism": "Baptized",
+          "baptismYes": "Yes, on {date}",
+          "baptismNo": "No",
+          "email": "Email",
+          "phone": "Phone",
+          "country": "Country",
+          "union": "Union",
+          "localField": "Local field",
+          "district": "District",
+          "church": "Church",
+          "club": "Club",
+          "section": "Section",
+          "internalId": "Internal ID",
+          "registrationDate": "Registration date",
+          "lastUpdate": "Last update",
+          "role": "Role",
+          "scopeType": "Scope type",
+          "scopeLocalFieldId": "Local field ID",
+          "scopeUnionId": "Union ID",
+          "relationship": "Relationship",
+          "bloodType": "Blood type",
+          "bloodUnspecified": "Not specified",
+          "allergies": "Allergies",
+          "diseases": "Conditions",
+          "medicines": "Medications"
+        },
+        "contacts": {
+          "principal": "Primary",
+          "call": "Call"
+        },
+        "access": {
+          "activeLabel": "Status",
+          "activeSub": "Enabled to sign in",
+          "appLabel": "Mobile app",
+          "appSub": "Access to mobile client",
+          "panelLabel": "Admin panel",
+          "panelSub": "Access to the admin panel"
+        }
       },
       "bulk": {
         "title": "Bulk user upload",
