+++ messages/es.json
@@ -11,7 +11,7 @@
       "members": "Miembros",
       "operations": "Operaciones",
       "rankings_analytics": "Rankings y análisis",
-      "materiales": "Materiales"
+      "materials": "Materiales"
     },
     "items": {
       "dashboard": "Dashboard",
@@ -104,10 +104,11 @@
       "subgroup_business": "Negocio",
       "subgroup_honors_cat": "Honores",
       "ranking_weights_root": "Configuración de pesos",
-      "materiales_bandeja": "Bandeja",
-      "materiales_comprobantes": "Comprobantes",
-      "materiales_inventario": "Inventario",
-      "materiales_configuracion": "Configuración"
+      "materials_inbox": "Bandeja",
+      "materials_receipts": "Comprobantes",
+      "materials_inventory": "Inventario",
+      "materials_categories": "Categorías",
+      "materials_config": "Configuración"
     },
     "breadcrumbs": {
       "dashboard": "Dashboard",
@@ -2911,16 +2912,126 @@
       },
       "detail": {
         "title": "Detalle de usuario",
+        "subtitle": "Información completa, accesos, salud y actividad del usuario.",
         "back": "Volver",
         "restrictedTitle": "Acceso restringido",
         "restrictedDescription": "Necesitas autorización global resuelta para consultar datos de terceros.",
         "statusActive": "Activo",
         "statusInactive": "Inactivo",
+        "ageYears": "{count, plural, one {# año} other {# años}}",
+        "approvalApproved": "Aprobado",
+        "approvalPending": "Pendiente",
+        "approvalRejected": "Rechazado",
+        "tabResumen": "Resumen",
+        "tabPersonalData": "Datos personales",
+        "tabHealth": "Salud",
+        "tabRolesAccess": "Roles y accesos",
         "tabInfo": "Información",
         "tabPostRegistration": "Post-registro",
         "tabSecurity": "Seguridad",
         "tabSessions": "Sesiones",
-        "postRegistrationUnavailable": "No se pudo obtener el estado del post-registro. El usuario puede no haber iniciado el proceso o puede haber un error de conectividad."
+        "postRegistrationUnavailable": "No se pudo obtener el estado del post-registro. El usuario puede no haber iniciado el proceso o puede haber un error de conectividad.",
+        "tenure": {
+          "years": "{count, plural, one {# año} other {# años}}",
+          "months": "{count, plural, one {# mes} other {# meses}}",
+          "days": "{count, plural, one {# día} other {# días}}"
+        },
+        "stats": {
+          "tenureLabel": "Tiempo en SACDIA",
+          "tenureRegisteredOn": "Registrado el {date}",
+          "classesLabel": "Clases",
+          "classesActive": "Inscripciones activas",
+          "rolesLabel": "Roles",
+          "rolesEmpty": "Sin roles",
+          "rolesRbacNote": "Vista de RBAC",
+          "lastUpdateLabel": "Última actualización",
+          "panelEnabled": "Panel habilitado",
+          "panelDisabled": "Panel deshabilitado",
+          "auditTracking": "Trazabilidad"
+        },
+        "sidebar": {
+          "title": "Acceso y estado",
+          "status": "Estado",
+          "approval": "Aprobación",
+          "roles": "Roles",
+          "appAccess": "Acceso a la app",
+          "panelAccess": "Acceso al panel",
+          "mfa": "MFA",
+          "assignments": "Asignaciones",
+          "enabled": "Habilitado",
+          "disabled": "Deshabilitado",
+          "yes": "Sí",
+          "no": "No",
+          "dash": "—"
+        },
+        "sections": {
+          "identityTitle": "Identidad",
+          "locationTitle": "Ubicación",
+          "fullDataTitle": "Datos completos",
+          "rolesAccessTitle": "Roles y accesos",
+          "scopeTitle": "Alcance",
+          "rolesLabel": "Roles activos",
+          "rolesEmpty": "El usuario no tiene roles asignados.",
+          "contactsTitle": "Contactos de emergencia",
+          "contactsEmpty": "Sin contactos de emergencia registrados.",
+          "contactsMissingPayload": "No se pudieron cargar los contactos.",
+          "legalRepTitle": "Representante legal",
+          "legalRepEmpty": "Sin representante legal registrado.",
+          "pastoralTitle": "Pastoral",
+          "pastoralEmpty": "Sin información pastoral registrada.",
+          "healthTitle": "Salud",
+          "healthBlockTitle": "Información médica",
+          "healthEmpty": "Sin información médica registrada.",
+          "healthShow": "Mostrar",
+          "healthHide": "Ocultar",
+          "healthProtectedTitle": "Información protegida",
+          "healthProtectedDescription": "Los datos médicos sólo se muestran a roles con permiso explícito."
+        },
+        "fields": {
+          "fullName": "Nombre completo",
+          "name": "Nombre",
+          "paternalLastName": "Apellido paterno",
+          "maternalLastName": "Apellido materno",
+          "birthday": "Fecha de nacimiento",
+          "gender": "Género",
+          "baptism": "Bautizado",
+          "baptismYes": "Sí, el {date}",
+          "baptismNo": "No",
+          "email": "Correo",
+          "phone": "Teléfono",
+          "country": "País",
+          "union": "Unión",
+          "localField": "Campo local",
+          "district": "Distrito",
+          "church": "Iglesia",
+          "club": "Club",
+          "section": "Sección",
+          "internalId": "ID interno",
+          "registrationDate": "Fecha de alta",
+          "lastUpdate": "Última actualización",
+          "role": "Rol",
+          "scopeType": "Tipo de alcance",
+          "scopeLocalFieldId": "ID de campo local",
+          "scopeUnionId": "ID de unión",
+          "relationship": "Parentesco",
+          "bloodType": "Tipo de sangre",
+          "bloodUnspecified": "Sin especificar",
+          "allergies": "Alergias",
+          "diseases": "Enfermedades",
+          "medicines": "Medicamentos"
+        },
+        "contacts": {
+          "principal": "Principal",
+          "call": "Llamar"
+        },
+        "access": {
+          "activeLabel": "Estado",
+          "activeSub": "Habilitado para iniciar sesión",
+          "appLabel": "App móvil",
+          "appSub": "Acceso al cliente móvil",
+          "panelLabel": "Panel admin",
+          "panelSub": "Acceso al panel administrativo"
+        }
       },
       "bulk": {
         "title": "Carga masiva de usuarios",
