+++ messages/fr.json
@@ -11,7 +11,7 @@
       "members": "Membres",
       "operations": "Opérations",
       "rankings_analytics": "Classements & Analytique",
-      "materiales": "Matériaux"
+      "materials": "Matériaux"
     },
     "items": {
       "dashboard": "Tableau de bord",
@@ -104,10 +104,11 @@
       "subgroup_business": "Affaires",
       "subgroup_honors_cat": "Honneurs",
       "ranking_weights_root": "Configuration des poids",
-      "materiales_bandeja": "Boîte de réception",
-      "materiales_comprobantes": "Reçus",
-      "materiales_inventario": "Inventaire",
-      "materiales_configuracion": "Paramètres"
+      "materials_inbox": "Boîte de réception",
+      "materials_receipts": "Reçus",
+      "materials_inventory": "Inventaire",
+      "materials_categories": "Catégories",
+      "materials_config": "Paramètres"
     },
     "breadcrumbs": {
       "dashboard": "Tableau de bord",
@@ -2833,17 +2834,127 @@
         "emptyDescription": "Essayez d’ajuster les filtres de recherche."
       },
       "detail": {
-        "title": "Détail de l’utilisateur",
+        "title": "Détail de l'utilisateur",
+        "subtitle": "Informations complètes, accès, santé et activité de l'utilisateur.",
         "back": "Retour",
         "restrictedTitle": "Accès restreint",
-        "restrictedDescription": "Vous avez besoin d’une autorisation globale résolue pour consulter les données d’autres utilisateurs.",
+        "restrictedDescription": "Vous avez besoin d'une autorisation globale résolue pour consulter les données d'autres utilisateurs.",
         "statusActive": "Actif",
         "statusInactive": "Inactif",
+        "ageYears": "{count, plural, one {# an} other {# ans}}",
+        "approvalApproved": "Approuvé",
+        "approvalPending": "En attente",
+        "approvalRejected": "Rejeté",
+        "tabResumen": "Résumé",
+        "tabPersonalData": "Données personnelles",
+        "tabHealth": "Santé",
+        "tabRolesAccess": "Rôles et accès",
         "tabInfo": "Informations",
         "tabPostRegistration": "Post-inscription",
         "tabSecurity": "Sécurité",
         "tabSessions": "Sessions",
-        "postRegistrationUnavailable": "Impossible de récupérer le statut post-inscription. L’utilisateur n’a peut-être pas commencé le processus ou il y a une erreur de connectivité."
+        "postRegistrationUnavailable": "Impossible de récupérer le statut post-inscription. L'utilisateur n'a peut-être pas commencé le processus ou il y a une erreur de connectivité.",
+        "tenure": {
+          "years": "{count, plural, one {# an} other {# ans}}",
+          "months": "{count, plural, one {# mois} other {# mois}}",
+          "days": "{count, plural, one {# jour} other {# jours}}"
+        },
+        "stats": {
+          "tenureLabel": "Temps dans SACDIA",
+          "tenureRegisteredOn": "Inscrit le {date}",
+          "classesLabel": "Classes",
+          "classesActive": "Inscriptions actives",
+          "rolesLabel": "Rôles",
+          "rolesEmpty": "Aucun rôle",
+          "rolesRbacNote": "Vue RBAC",
+          "lastUpdateLabel": "Dernière mise à jour",
+          "panelEnabled": "Panneau activé",
+          "panelDisabled": "Panneau désactivé",
+          "auditTracking": "Traçabilité"
+        },
+        "sidebar": {
+          "title": "Accès et statut",
+          "status": "Statut",
+          "approval": "Approbation",
+          "roles": "Rôles",
+          "appAccess": "Accès à l'app",
+          "panelAccess": "Accès au panneau",
+          "mfa": "MFA",
+          "assignments": "Affectations",
+          "enabled": "Activé",
+          "disabled": "Désactivé",
+          "yes": "Oui",
+          "no": "Non",
+          "dash": "—"
+        },
+        "sections": {
+          "identityTitle": "Identité",
+          "locationTitle": "Localisation",
+          "fullDataTitle": "Données complètes",
+          "rolesAccessTitle": "Rôles et accès",
+          "scopeTitle": "Portée",
+          "rolesLabel": "Rôles actifs",
+          "rolesEmpty": "Cet utilisateur n'a aucun rôle assigné.",
+          "contactsTitle": "Contacts d'urgence",
+          "contactsEmpty": "Aucun contact d'urgence enregistré.",
+          "contactsMissingPayload": "Impossible de charger les contacts.",
+          "legalRepTitle": "Représentant légal",
+          "legalRepEmpty": "Aucun représentant légal enregistré.",
+          "pastoralTitle": "Pastoral",
+          "pastoralEmpty": "Aucune information pastorale enregistrée.",
+          "healthTitle": "Santé",
+          "healthBlockTitle": "Informations médicales",
+          "healthEmpty": "Aucune information médicale enregistrée.",
+          "healthShow": "Afficher",
+          "healthHide": "Masquer",
+          "healthProtectedTitle": "Informations protégées",
+          "healthProtectedDescription": "Les données médicales ne sont affichées qu'aux rôles disposant d'une permission explicite."
+        },
+        "fields": {
+          "fullName": "Nom complet",
+          "name": "Prénom",
+          "paternalLastName": "Nom paternel",
+          "maternalLastName": "Nom maternel",
+          "birthday": "Date de naissance",
+          "gender": "Genre",
+          "baptism": "Baptisé",
+          "baptismYes": "Oui, le {date}",
+          "baptismNo": "Non",
+          "email": "Email",
+          "phone": "Téléphone",
+          "country": "Pays",
+          "union": "Union",
+          "localField": "Champ local",
+          "district": "District",
+          "church": "Église",
+          "club": "Club",
+          "section": "Section",
+          "internalId": "ID interne",
+          "registrationDate": "Date d'inscription",
+          "lastUpdate": "Dernière mise à jour",
+          "role": "Rôle",
+          "scopeType": "Type de portée",
+          "scopeLocalFieldId": "ID de champ local",
+          "scopeUnionId": "ID d'union",
+          "relationship": "Lien de parenté",
+          "bloodType": "Groupe sanguin",
+          "bloodUnspecified": "Non spécifié",
+          "allergies": "Allergies",
+          "diseases": "Maladies",
+          "medicines": "Médicaments"
+        },
+        "contacts": {
+          "principal": "Principal",
+          "call": "Appeler"
+        },
+        "access": {
+          "activeLabel": "Statut",
+          "activeSub": "Activé pour se connecter",
+          "appLabel": "App mobile",
+          "appSub": "Accès au client mobile",
+          "panelLabel": "Panneau admin",
+          "panelSub": "Accès au panneau administratif"
+        }
       }
     }
   },
