+++ messages/pt-BR.json
@@ -11,7 +11,7 @@
       "members": "Membros",
       "operations": "Operações",
       "rankings_analytics": "Rankings e análises",
-      "materiales": "Materiais"
+      "materials": "Materiais"
     },
     "items": {
       "dashboard": "Dashboard",
@@ -104,10 +104,11 @@
       "subgroup_business": "Negócio",
       "subgroup_honors_cat": "Honras",
       "ranking_weights_root": "Configuração de pesos",
-      "materiales_bandeja": "Caixa de entrada",
-      "materiales_comprobantes": "Comprovantes",
-      "materiales_inventario": "Inventário",
-      "materiales_configuracion": "Configurações"
+      "materials_inbox": "Caixa de entrada",
+      "materials_receipts": "Comprovantes",
+      "materials_inventory": "Inventário",
+      "materials_categories": "Categorias",
+      "materials_config": "Configurações"
     },
     "breadcrumbs": {
       "dashboard": "Painel",
@@ -2833,17 +2834,127 @@
         "emptyDescription": "Tente ajustar os filtros de pesquisa."
       },
       "detail": {
-        "title": "Detalhes do usuário",
+        "title": "Detalhe do usuário",
+        "subtitle": "Informações completas, acessos, saúde e atividade do usuário.",
         "back": "Voltar",
         "restrictedTitle": "Acesso restrito",
         "restrictedDescription": "Você precisa de autorização global resolvida para consultar dados de outros usuários.",
         "statusActive": "Ativo",
         "statusInactive": "Inativo",
+        "ageYears": "{count, plural, one {# ano} other {# anos}}",
+        "approvalApproved": "Aprovado",
+        "approvalPending": "Pendente",
+        "approvalRejected": "Rejeitado",
+        "tabResumen": "Resumo",
+        "tabPersonalData": "Dados pessoais",
+        "tabHealth": "Saúde",
+        "tabRolesAccess": "Perfis e acessos",
         "tabInfo": "Informações",
         "tabPostRegistration": "Pós-registro",
         "tabSecurity": "Segurança",
         "tabSessions": "Sessões",
-        "postRegistrationUnavailable": "Não foi possível obter o status do pós-registro. O usuário pode não ter iniciado o processo ou pode haver um erro de conectividade."
+        "postRegistrationUnavailable": "Não foi possível obter o status do pós-registro. O usuário pode não ter iniciado o processo ou pode haver um erro de conectividade.",
+        "tenure": {
+          "years": "{count, plural, one {# ano} other {# anos}}",
+          "months": "{count, plural, one {# mês} other {# meses}}",
+          "days": "{count, plural, one {# dia} other {# dias}}"
+        },
+        "stats": {
+          "tenureLabel": "Tempo no SACDIA",
+          "tenureRegisteredOn": "Registrado em {date}",
+          "classesLabel": "Turmas",
+          "classesActive": "Matrículas ativas",
+          "rolesLabel": "Perfis",
+          "rolesEmpty": "Sem perfis",
+          "rolesRbacNote": "Visão RBAC",
+          "lastUpdateLabel": "Última atualização",
+          "panelEnabled": "Painel habilitado",
+          "panelDisabled": "Painel desabilitado",
+          "auditTracking": "Rastreabilidade"
+        },
+        "sidebar": {
+          "title": "Acesso e status",
+          "status": "Status",
+          "approval": "Aprovação",
+          "roles": "Perfis",
+          "appAccess": "Acesso ao app",
+          "panelAccess": "Acesso ao painel",
+          "mfa": "MFA",
+          "assignments": "Atribuições",
+          "enabled": "Habilitado",
+          "disabled": "Desabilitado",
+          "yes": "Sim",
+          "no": "Não",
+          "dash": "—"
+        },
+        "sections": {
+          "identityTitle": "Identidade",
+          "locationTitle": "Localização",
+          "fullDataTitle": "Dados completos",
+          "rolesAccessTitle": "Perfis e acessos",
+          "scopeTitle": "Escopo",
+          "rolesLabel": "Perfis ativos",
+          "rolesEmpty": "Este usuário não tem perfis atribuídos.",
+          "contactsTitle": "Contatos de emergência",
+          "contactsEmpty": "Nenhum contato de emergência registrado.",
+          "contactsMissingPayload": "Não foi possível carregar os contatos.",
+          "legalRepTitle": "Representante legal",
+          "legalRepEmpty": "Nenhum representante legal registrado.",
+          "pastoralTitle": "Pastoral",
+          "pastoralEmpty": "Nenhuma informação pastoral registrada.",
+          "healthTitle": "Saúde",
+          "healthBlockTitle": "Informações médicas",
+          "healthEmpty": "Nenhuma informação médica registrada.",
+          "healthShow": "Mostrar",
+          "healthHide": "Ocultar",
+          "healthProtectedTitle": "Informações protegidas",
+          "healthProtectedDescription": "Os dados médicos são exibidos apenas para perfis com permissão explícita."
+        },
+        "fields": {
+          "fullName": "Nome completo",
+          "name": "Nome",
+          "paternalLastName": "Sobrenome paterno",
+          "maternalLastName": "Sobrenome materno",
+          "birthday": "Data de nascimento",
+          "gender": "Gênero",
+          "baptism": "Batizado",
+          "baptismYes": "Sim, em {date}",
+          "baptismNo": "Não",
+          "email": "E-mail",
+          "phone": "Telefone",
+          "country": "País",
+          "union": "União",
+          "localField": "Campo local",
+          "district": "Distrito",
+          "church": "Igreja",
+          "club": "Clube",
+          "section": "Seção",
+          "internalId": "ID interno",
+          "registrationDate": "Data de cadastro",
+          "lastUpdate": "Última atualização",
+          "role": "Perfil",
+          "scopeType": "Tipo de escopo",
+          "scopeLocalFieldId": "ID de campo local",
+          "scopeUnionId": "ID de união",
+          "relationship": "Parentesco",
+          "bloodType": "Tipo sanguíneo",
+          "bloodUnspecified": "Não especificado",
+          "allergies": "Alergias",
+          "diseases": "Doenças",
+          "medicines": "Medicamentos"
+        },
+        "contacts": {
+          "principal": "Principal",
+          "call": "Ligar"
+        },
+        "access": {
+          "activeLabel": "Status",
+          "activeSub": "Habilitado para fazer login",
+          "appLabel": "App móvel",
+          "appSub": "Acesso ao cliente móvel",
+          "panelLabel": "Painel admin",
+          "panelSub": "Acesso ao painel administrativo"
+        }
       }
     }
   },
