--- src/app/(dashboard)/dashboard/materiales/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function MaterialesIndexPage() {
-  redirect("/dashboard/materiales/bandeja");
-}
