+++ src/app/(dashboard)/dashboard/materials/page.tsx
@@ -1,0 +1,5 @@
+import { redirect } from "next/navigation";
+
+export default function MaterialesIndexPage() {
+  redirect("/dashboard/materials/inbox");
+}
