+++ src/app/(dashboard)/dashboard/materials/receipts/_components/receipts-filters.tsx
@@ -1,0 +1,54 @@
+"use client";
+
+import { useRouter, useSearchParams, usePathname } from "next/navigation";
+import { useCallback, useTransition } from "react";
+import { Search } from "lucide-react";
+import { Input } from "@/components/ui/input";
+
+interface ReceiptsFiltersProps {
+  currentQ: string;
+}
+
+export function ReceiptsFilters({ currentQ }: ReceiptsFiltersProps) {
+  const router = useRouter();
+  const pathname = usePathname();
+  const searchParams = useSearchParams();
+  const [, startTransition] = useTransition();
+
+  const updateParam = useCallback(
+    (key: string, value: string | undefined) => {
+      const params = new URLSearchParams(searchParams.toString());
+      if (value) {
+        params.set(key, value);
+      } else {
+        params.delete(key);
+      }
+      // Reset to page 1 on filter change
+      params.delete("page");
+      startTransition(() => {
+        router.push(`${pathname}?${params.toString()}`);
+      });
+    },
+    [router, pathname, searchParams],
+  );
+
+  function handleSearchChange(e: React.ChangeEvent<HTMLInputElement>) {
+    const value = e.target.value.trim();
+    updateParam("q", value || undefined);
+  }
+
+  return (
+    <div className="flex flex-wrap items-center gap-3">
+      {/* Search input */}
+      <div className="relative">
+        <Search className="absolute left-2.5 top-1/2 size-4 -translate-y-1/2 text-muted-foreground" />
+        <Input
+          className="h-9 w-[260px] pl-8"
+          placeholder="Buscar folio o director..."
+          defaultValue={currentQ}
+          onChange={handleSearchChange}
+        />
+      </div>
+    </div>
+  );
+}
