+++ src/i18n/messages.d.ts
@@ -2880,16 +2880,126 @@ export interface IntlMessages {
       };
       detail: {
         title: string;
+        subtitle: string;
         back: string;
         restrictedTitle: string;
         restrictedDescription: string;
         statusActive: string;
         statusInactive: string;
+        ageYears: string;
+        approvalApproved: string;
+        approvalPending: string;
+        approvalRejected: string;
+        tabResumen: string;
+        tabPersonalData: string;
+        tabHealth: string;
+        tabRolesAccess: string;
         tabInfo: string;
         tabPostRegistration: string;
         tabSecurity: string;
         tabSessions: string;
         postRegistrationUnavailable: string;
+        tenure: {
+          years: string;
+          months: string;
+          days: string;
+        };
+        stats: {
+          tenureLabel: string;
+          tenureRegisteredOn: string;
+          classesLabel: string;
+          classesActive: string;
+          rolesLabel: string;
+          rolesEmpty: string;
+          rolesRbacNote: string;
+          lastUpdateLabel: string;
+          panelEnabled: string;
+          panelDisabled: string;
+          auditTracking: string;
+        };
+        sidebar: {
+          title: string;
+          status: string;
+          approval: string;
+          roles: string;
+          appAccess: string;
+          panelAccess: string;
+          mfa: string;
+          assignments: string;
+          enabled: string;
+          disabled: string;
+          yes: string;
+          no: string;
+          dash: string;
+        };
+        sections: {
+          identityTitle: string;
+          locationTitle: string;
+          fullDataTitle: string;
+          rolesAccessTitle: string;
+          scopeTitle: string;
+          rolesLabel: string;
+          rolesEmpty: string;
+          contactsTitle: string;
+          contactsEmpty: string;
+          contactsMissingPayload: string;
+          legalRepTitle: string;
+          legalRepEmpty: string;
+          pastoralTitle: string;
+          pastoralEmpty: string;
+          healthTitle: string;
+          healthBlockTitle: string;
+          healthEmpty: string;
+          healthShow: string;
+          healthHide: string;
+          healthProtectedTitle: string;
+          healthProtectedDescription: string;
+        };
+        fields: {
+          fullName: string;
+          name: string;
+          paternalLastName: string;
+          maternalLastName: string;
+          birthday: string;
+          gender: string;
+          baptism: string;
+          baptismYes: string;
+          baptismNo: string;
+          email: string;
+          phone: string;
+          country: string;
+          union: string;
+          localField: string;
+          district: string;
+          church: string;
+          club: string;
+          section: string;
+          internalId: string;
+          registrationDate: string;
+          lastUpdate: string;
+          role: string;
+          scopeType: string;
+          scopeLocalFieldId: string;
+          scopeUnionId: string;
+          relationship: string;
+          bloodType: string;
+          bloodUnspecified: string;
+          allergies: string;
+          diseases: string;
+          medicines: string;
+        };
+        contacts: {
+          principal: string;
+          call: string;
+        };
+        access: {
+          activeLabel: string;
+          activeSub: string;
+          appLabel: string;
+          appSub: string;
+          panelLabel: string;
+          panelSub: string;
+        };
       };
       bulk: {
         title: string;
